lib: add sourceFilesByManifest{,JsonFile} filter methods
Added sourceFilesByManifest* filter methods to allow the filtering of
source files by explicit manifest. This is particularly useful for using
using the output of `git ls-files` to provide to a derivation only the
files that would be present in a fresh clone (as performed by Hydra, for
example), and also for preventing large or sensitive files from being
imported into the store.