glibc: configure default 'zonedir' for Nix zoneinfo
This patch updates glibc to encode a default path for locating tzdata
in the absence of an overriding TZDIR environment variable. This allows
Nix packages to continue working in situations where it is not possible
to reliably configure the environment. For example, this protects
against breakage that can occur with the use of `env -i`.