glibc: add symlink to add ldconfig functionality
ldconfig(8) is the standard Linux mechanism for configuring "system"
default libraries without resorting to the use of environment variables.
This mechanism is useful too in Nix, particularly as a way to provide
access to certain nss libraries for use with PAM (e.g. nss_ldap) without
the use of environment variables.

This patch restores the ldconfig functionality by updating glibc to
reference ld.so.cache in /nix/etc/ld.so.cache. We are proposing the use
of this path to be symmetric with /nix/var, and as a location that an
administrator of the system can manage in any number of ways.