glibc: add "system" location to nix-locale-archive.patch
Update locale search logic to check a default Nix "system" path after
checking the LOCALE_ARCHIVE* environment variables and before reverting
the glibc default.

This patch also adds the LOCALE_ARCHIVE{,_2_27} variables to the list of
"unsecure" variables to be pared from the environment prior to invoking
setuid/setgid programs as defined in `unsecvars.h`. This is consistent
with glibc's treatment of LOCPATH which serves the same purpose outside
of Nix.