TEMPORARY: revert flox changes to use upstream cache
The patches on the flox fork of nixpkgs solve important problems for
using Nix packages on non-NixOS distros (and in situations where you
have little or no control over environment variables in general), but
they are not necessary for demonstrating the Flox value prop at this
stage. Adding a single commit which counteracts the previous four for
the purpose of getting a complete copy of nixpkgs artifacts from
nixos.org in the least amount of time.